+++ packages/atomic/vitest-utils/setup.ts
@@ -1,7 +1,11 @@
-import {afterEach} from 'vitest';
-import {fixtureCleanup} from './testing-helpers/fixture-wrapper';
+import {beforeEach} from 'vitest';
+import '../src/themes/coveo.css';
+import '../src/utils/coveo.tw.css';
+import '../src/utils/tailwind-utilities/utilities.tw.css';
+import '../src/utils/tailwind.global.tw.css';
+import {fixtureCleanup} from './testing-helpers/fixture-wrapper.js';
 
-afterEach(async () => {
+beforeEach(async () => {
   document.adoptedStyleSheets = [];
   fixtureCleanup();
 });